requirements.txt
beautifulsoup4==4.12.2
newspaper3k==0.2.8
requests==2.31.0
tqdm==4.66.1
